Moved FromBridgedChainMessagesDeliveryProof to bp-messages (#2169)
* moved FromBridgedChainMessagesDeliveryProof to bp-messages

* spelling

* fix benchmarks compilation
